+++ zcash_client_memory/src/types/transparent.rs
@@ -7,6 +7,7 @@ use ::transparent::{
     address::TransparentAddress,
     bundle::{OutPoint, TxOut},
 };
+use transparent::keys::TransparentKeyScope;
 use zcash_client_backend::wallet::WalletTransparentOutput;
 
 use zcash_protocol::{consensus::BlockHeight, TxId};
@@ -88,6 +89,8 @@ pub struct ReceivedTransparentOutput {
     pub(crate) account_id: AccountId,
     // The address to which this TXO was sent
     pub(crate) address: TransparentAddress,
+    // The key scope at which the address was derived
+    pub(crate) key_scope: TransparentKeyScope,
     // script, value_zat
     pub(crate) txout: TxOut,
     /// The maximum block height at which this TXO was either
@@ -103,6 +106,7 @@ impl ReceivedTransparentOutput {
         transaction_id: TxId,
         account_id: AccountId,
         address: TransparentAddress,
+        key_scope: TransparentKeyScope,
         txout: TxOut,
         max_observed_unspent_height: BlockHeight,
     ) -> Self {
@@ -110,6 +114,7 @@ impl ReceivedTransparentOutput {
             transaction_id,
             account_id,
             address,
+            key_scope,
             txout,
             max_observed_unspent_height: Some(max_observed_unspent_height),
         }
@@ -171,6 +176,10 @@ mod serialization {
         }
     }
 
+    // FIXME: Key scope information needs to be added to both `proto::Address` and
+    // `proto::ReceivedTransparentOutput`, with a data migration that updates stored data with
+    // correct scope information.
+    #[allow(unreachable_code)]
     impl TryFrom<proto::ReceivedTransparentOutput> for ReceivedTransparentOutput {
         type Error = crate::Error;
 
@@ -179,6 +188,7 @@ mod serialization {
                 transaction_id: TxId::from_bytes(output.transaction_id.clone().try_into()?),
                 account_id: output.account_id.into(),
                 address: TransparentAddress::decode(&EncodingParams, &output.address)?,
+                key_scope: todo!(),
                 txout: read_optional!(output, txout)?.try_into()?,
                 max_observed_unspent_height: output.max_observed_unspent_height.map(|h| h.into()),
             })